src/requirements-dev.txt
-r requirements.txt
black==24.4.2
pytest==8.2.2
pytest-django==4.8.0
responses==0.25.3